Start to organize the code better
Move goroutines out of the network layer and into the STM layer (still
wrong, but better for now). Make the data layer expose gConn as a
ReaderWriter by modifying method signatures to implement the
interface. This will make it more composable.

Teach the server code to decode the protos sent by the client and
handle keystrokes, proxying them to a new pty object that should be
running a login shell.