We have basic bidirectional communication now
We don't handle EOF from the server and we're not doing any actual
state managment yet. But keystrokes from the client make it to the
server and interact with the shell there. Output from the shell is
mirrored back. The next step will be to avoid reading from the remote
shell process until the client is connected so that we actually
display a prompt initially.