Streamline STM to have a single Run() and Shutdown()
Use the same Run() and Shutdown() for both client and server,
differentiating the things that aren't the same based on a property of
the object. This lets us have a simpler overall interface for now. We
also get a chance to cleanup and standardize the shutdown handling so
it's better.