Update the Fragment message in our tranport proto
Now that we're going to build that functionality out, we're thinking a
bit more about it. We need an id, but that's for a group of
fragments. Each actual fragment should include it's sequence within
the group and the total expected fragments. We also allow for both
compressed and uncompressed fragments.