Convert nonce to a proper object
This should reduce errors in using it as before it was directly
available in the upstream objects that used a nonce and they could
maniuplate it. We couldn't safely wrap it in a mutex that way, either
and that's important for security here because gosh is using multiple
goroutines and could access the nonce from more than one.