Rename the proto file and go module
This proto will encapsulate all messages we use in gosh and isn't
transport specific. Until now, everything we'd defined was going to be
used in the network layer and thus transport was an ok name. Now,
we're going to do fragmentation in the state management layer though,
so a more general name is better. s/transport/goshpb/g.