Change parser interface
Have calls to parse* in the parser return a list of actions instead of
relying on a callback. This should make it a more straight forward
call flow. It will work as long as we don't have combinations of
things that require cleaning up params or intermediate data between
usage. That shouldn't be the case, so this should be ok. There are
still some ugly warts here, around how we handle the difference
between bytes and runes. Clean that up later, as we iterate.