Change the way we run the terminal to setup for adding diffs
We want to start reading diffs from the terminal state and shipping
those intstead of just writing to the network and then to stdout on
the client. To do this, we pass in the pty reader object for the
terminal and teach the terminal how to Run(). In Run, we read from the
reader object until eof. On the client side, we mock this up with pipe
object so the read from the network can be written into the pipe,
which the terminal reads.