Add cell.diff() which builds on format.diff()
Add tests. This diff is a complete diff, but we'll also augment with a
version that leverages having a format passed in so we can more
efficiently change the pen attributes as we build up a framebuffer
diff. This will save spitting out the same format diff for consecutive
cells with the same format.