Teach framebuffer how to dump an OSC string that defines the size
This isn't a standard thing in terminals, but we want a way for one
framebuffer to be able to communicate the size it expects another
framebuffer to be. To do this, we'll hijack "X" as an OSC command and
allow rows and columns as paramters. Doing this, we can then print our
own OSC string to communicate to a peer.