Move the top/bottom/left/right margin values to Terminal
These shouldn't be properties of the framebuffer (dumb), but of the
terminal (smart). Move them out of framebuffer which will simplify
that object and into Terminal where it'll make more sense for them. As
we build out margin support, we'll need to improve this further, but
for now, at least they're in the right place.