Use a margin object to hold top/bottom/left/right margin values
Using an object will let us query as to whether the margin is set or
not which will be useful for determining diffs and when we begin using
it to change motion, scrolling, etc.

We also move the cursor to the top/left of the page when we set
margins to match DEC behaviour.