Get the Terminal.TestPrint function ready for margins
Make it easier to adapt this test to be margin aware by adding a
helper function that creates a terminal with cursor and fb
setup. We'll extend this to provide a version that also wants a margin
so we can then add the tests needed to validate behaviour when a
margin is in play.