Add framebuffer.getRegion() which returns a view from the data
By adding this method, we can change how we scroll a frambebuffer and
then have a clean abstraction for scrolling either the full
framebuffer or a portion of it. Because we return the view as a
framebuffer object, we can use all of the normal methods like getting
and setting cells, etc.