Fix bug in server shutdown
We were calling a method on a nil object on the server side and
forcing a crash, which looked like it was doing what we wanted, but
was masking an unclean shutdown. Instead, use the cancelPty() method
and rely on the terminal to recognize the read error from the now
closed pty to terminate the read loop.