Abstract knowledge of the network away from STM
Make stm take a ReadWriteCloser and call it remote. We don't need to
care about what's behind the remote and this helps us decouple
things. To make this work, we had to make the Close method on GConn
return an error to match the expected interface. We now log errors
from closing the remote as well.