Teach stmObj how to read and write in fragments
When sending messages, fragment them first so we're always shipping
goshpb.Fragments around. When reading, store the messages in the
Fragmenter and only consume them if they're complete.

By doing this, we're able to send diffs that exceed the message sizes
we're trying to work within to keep overall packet sizes within an
expected bound.