Refactor the main gosh client binary a bit
Move the code for running the remote server into a separate function
that returns the connection data. This makes reading the main() method
a bit nicer as it focuses more clearly on what we're doing.

Do the same for the client side. That's less important on its own, but
does make it look the same as the server setup code and it gives us a
chance to nicely use the connection data from the server process with
appropriate names.