Generalize private flags into an object and carry them in a map
We're going to add more csi private flags and the machinery for
handling them was a bit clumsy. Turn them into objects that can
compare themselves, return their ANSI string, etc. With this approach,
adding a new one is much simpler and will always be included in the
diff() machinery without further modification.