Handle ESC M
This command moves the cursor up a row and will scroll the screen down
if at the top already.

The less command depends on this behaviour when scrolling back through
it's buffer as it moves to 0,0, sends ESC M and then prints the row,
iterating through the rows it wants to display in reverse order.