Only print ; separator for H cursor ansi if we print the column
If we print a row and no column, we don't need ';' to separate the
parameters. If we print a column, we always need a ';' to ensure the
number we print is recognized as the second parameter and thus
column. Overall, a minor simplification but it's more consistent with
the ansi generation from other tools.