Track last change time in the terminal object
Using this, we can short circuit Diff() comparisons cheaply as if the
time hasn't changed, nothing has altered the object. This requires
every path that changes state to update the timestamp.

The timestamp will also be useful as we start to build out the needed
state management where we do diffs against the last "acked" object
from the remote side instead of the last snapshot we took.