Handle CSI t - xterm/dtterm winops
tmux uses this to ask the terminal to save and restore the title and
icon, based on an xterm feature. We'll honor that, but without a stack
to save multiple items. If we discover this is problematic later, we
can easily add a stack, but for now we expect this is pushed and the
popped at a depth of 1 so we don't need the complexity.