Split out cursor movement methods
Instead of a giant cursorMove() method, add individual methods per
movement type. We put these in a separate file. We still use
cursorMove() as the umbrella method for calling them, so this is
transparent to the upstream users for now. The idea behind splitting
them out is that we're going to start adding more sophistication to
some of them as we work to pass vvtest. Things like handling reverse
wrap will mean that the case statement blocks were going to get large
and testing was going to get uglier. By splitting things out, we can
target tests more effectively.