Drop some unused or ignorable SGR codes
Notable in this removal/cleanup is that we will drop support for
double underline, italic and faint brightness as style
attributes. XTerm doesn't support these, so we should consider it so
legacy that nobody will miss it. That lets us simplify how we handle
styles as well because we no longer need different levels of
brightness or underline. At this point, we could move to storing
attributes in a bitfield int, but we'll save that for a followup.