Change format attribute setting to use a plain function
Because we aren't using pointers to format objects, the setAttr had to
return the updated format object. We could have changed this method to
be a pointer receiver and used a format pointer when calling it, but
that was a bit ugly too. Instead, just do setAttr outside of any
object, pass the current attr value in and record the result.