Fix cursor handling in Terminal.Diff
We need to spit out a CUP command if either we found differences in
the framebuffer (which moved us around the screen to regenerate) or if
the terminal are otherwise the same but the cursors are in different
locations. Add a test that catches the latter case as that was missed
previously.