Correct a margin related bug introduced in ebe8a1
When we started copying margins in Terminal.Copy(), we started to ship
those as part of our diff and that breaks client side handling. This
has clarified our understanding of what we need to ship in diffs. For
the most part, this will be only the visual differences and modes that
may affect visual differences. We don't need to ship every mode unless
it is something like the REVERSE VIDEO mode that changes how the
client side renderer works. This change only removes the handling of
margins with Diff() but we'll followup with changes that only ship
relevant modes to the client as well.