Rename several functions
We want to move away from getFoo to foo and to be more consistent with
naming things that return ansi sequences, etc. This is leading up to
some other refactoring to be more consistent about how we query
whether we're in region or what the margins are when making decisions
about handling various commands.