Add some new boundedMarginFoo helper methods
We had a patter of testing for being in the region and then adjusting
a left, right, top or bottom value accordingly, so centralize that
logic to reduce potential errors. This should also be nicer than
inScrollingRegion because it only cares about a particular dimension.

Make the cursorUp/Down/Forward/Back methods use the new functionality
first. Other conversions to follow.