Change message from error to debug
We don't want to dump this in logs except for debugging as it
shouldn't happen in reality. We can increase log verbosity to see it
if it becomes a real problem in the wild. Save anything higher than
debug for stuff the user would be interested about like remote ip
changes, etc.