Add setRow and setCol methods
This lets us avoid manipulating the cursor directly outside of
dedicated methods. We can't use cursorMoveAbs in this case because it
always bounds the cursor within the framebuffer dimensions. When
printing, it is ok to have the cursor take a position outside the
border though, which is the sentinel that we need to consider
autowrap, etc.