Change mode handling again
In the previous implementation, it was pretty ugly to determine mode
status at runtime. Revamp how we store modes, moving to a
concatenation of the data string and the mode id as the key, with a
single human friendly name mapping to those unfriendly identifiers. We
change internal representation of mode as well. With this, we can
query for a well known name (string) and get state.

We also facilitate "enabled" (SET) by default mechanics as newMode now
takes an initial state. We noticed the need for this when using vttest
with LNM (newline/linefeed) mode test. That test still doesn't pass,
but we should be to make it do so more easily after this change.