Only ship diffs for modes when they indicate shouldTransport()
Most modes affect how the rendering is done and don't need to actually
be shipped to the client. Teach modes how to indicate if they should
be shipped or not. Going forward, we will only ship modes that
materially affect the display on the client side. This includes things
like reverse video, which need the client side terminal software to do
something on behalf of the emulation. We have setup a few modes to be
tranpsorted by default, but this may not be complete or accurate yet.