Clean up handling of mode transporting in diffs
Don't use a slice presized to the number of modes, but give it that
much capacity, just in case. Then, append to the slice instead of
inserting at a specific location. This avoids empty id's in the array
when we actually go to look for them as we build the diff.

We should think about just predefining the list of transportable modes
and not building this every time.