Fix how we find the combinable char on the previous row
We can't just assume the last column of the previous row is where the
combinable character is. We need to work back from the end to the last
cell of the row that is actually set.

Add a test case to cover this.

With this, we now also handle the `echo -e "abc\n\0314\0202\ndef"`
case that mosh notes.