Teach parser and parameters objects to copy() themselves
When we copy a terminal, we're also going to copy the parser and that
means we need to copy parameters too. We want to ensure that when we
write into a terminal to apply a diff, we're always writing into the
same parser state rather than one that may have been part way through
processing something else.