Add new capabilities to Terminal objects
When we snapshot in the state management protocol, we'll ask the
server side terminal for it's last change timestamp and use that as an
identifier. Thus, we add Lastchange to return that value.

We also want to replace the terminal on the client side with the
canondical version after applying target to source. This ensures that
goroutines we spin up at Run() time are all working with the correct
object still.