Add state management
When the server sends a diff, it indicates the source and target
timestamp. The client will use these timestamps to apply the diff.

If the client thinks the target timestamp is already older than it's
current view, it will ignore the message. This protects against out of
order packets. If we can't find the source, we also ignore the
message.

If the above two conditions are met, we apply the diff from the server
to the source terminal and then make that terminal live. If the source
we applied to is older than what we think our current view represents,
we also generate a local diff to bring it up to current.