Teach fragmenter how to cleanup orphaned fragments
If we experience loss of fragments and are not able to fully assemble
a set, we'd just build up memory usage indefinitely. To combat this,
have Fragger how to track last seen timestamps for a fragment
set. From that, we add a Clean method that will drop partial sets over
1 minute old. Clients can use this to ensure good hygiene.