Leave == on our base64 encoded key
Most strips == from the base64 encoded key before dumping it to stdout
and then puts it back on when consuming it from the environment on the
client process. I suspect, but don't know for sure, that this was done
because many terminals consider '=' to be a non-word character and
thus selecting the key with the mouse by double clicking on it would
omit the "==" at the end of the encoded key from being selected. That
was likely a hack early on and just lived forever. We'll stop
propagating that hack and dump the natural encoding and expect to
consume an natural encoding.