Handle username@hostname argument
Rename argument host to dest and accept {username@}hostname instead.

The ssh cli has always accepted {username@}hostname with the optional
username portion. We want to mimick that capability but make a first
step toward it. This change allows us to specify username@hostname
where we will connect to hostname with gosh-client after running the
server, but specify username@hostname to the initial ssh connection.

A follow up step will try to relax argument parsing so we can just say
`gosh username@hostname` or `gosh hostname` instead of requiring the
--dest flag.