Make shutdown a bit clean, add logging
Shutdown triggers a remote shutdown message, which would in turn see
the recipient send us a shutdown message again. Try to avoid multiple
attempts to shutdown by skipping it if our own shutdown variable is
true. There is likely a race here, but it's ok because there isn't any
harm if this runs more than once, it's just reducing noise in the logs.