Change the way we handle the ssh auth sockets
Have the client and server binaries create their sockets and take
responsibility for closing them down in main(). Pass the respective
objects into the stmObj constructor.

On the server side, use the pid of the process in the name of the
socket file, so that failure to terminate cleanly doesn't prevent
starting the next process where agent forwarding is requested. This
also facilitates multiple sockets for different gosh instances without
trampling each other.