Add a new enum to the proto - heartbeat, heartbeat_ack
We're going to wire up support for detecting when we need to overlay a
banner on the local terminal. That means we'll need to send some
traffic when the server isn't so we can detect whether it's
gone (which could include a network break) or just not sending diffs
because it doesn't need to.

Add the enum values, renumber others. These won't have payload at all,
not even a timestamp. Instead, we'll simply use them as triggers to
update the time we last saw the remote.