Allow heartbeat backoff
Have the sendPayload method return an error if writing to the remote
peer failed. Use that as a signal in heartbeat to back off, to a
maximum of 64 seconds.

This will only help when the local writes fail. If we send packets
successfully, but they're not reciprocated by the peer, we'll still
use the 1s interval. We can adjust this later though, if needed.