Prevent flashing of the overlay
Only display the overlay if it's been 1.5 of the stale interval. This
means that we get a chance to send our heartbeat message and get the
ack to it in 30s. Doing this means that we won't flash the banner and
then immediately remove it when the server is fine but the session is
just quiet.