Change nonce handling
We had restricted ourselves to 2^32-1 (we never use 0) nonce IVs being
generated on client and server based on a similar restriction in mosh,
which used OCB3 instead of GCM. With GCM, in counter (rather than
random) mode, it should be safe to use 2^64 IVs, so make more use of
the IV space. We should never need to worry about exhausting this for
a single session. Even with the previous approach, we had lots of
room, but this should remove any thinking about ever overflowing it.

To achieve this change, we do the following:
* Continue to avoid using 0 on either side
* Continue using 1 bit of the 12 bytes for defining "server"
  vs "client", and treat this one byte part of the 2^64 bit IV space,
  which makes 1 bit of the 2^64-1 constant for each side, dividing the
  64-bit space in half between client and server
* Allow each side of the connection to use up to 2^62-1 values.