Add OSC8 hyperlink handling
To be able to ship diffs that include sensible data about OSC8
hyperlinks, each cell will need to have some hyperlink
information. We'll attach that to cell objects, defaulting every cell
to carry "8;;" (end hyperlink).

We modify a few helper methods and extend the extendedDiff helper to
be able to spit out sensible diffs.

When we save and restore the cursor, we consider the current osc8 data
as part of that, just as we do the charset. We reset the data when the
terminal is reset and copy it around as needed.