Change Terminal.Copy semantics
We want to reduce the number of objects created unnecessarily, forcing
a lot of work for the garbage collector. So, make Copy into copy and
provide two new external methods for clients to use. One will only
return the copy if the provided timestamp is earlier than the last
change state of this terminal. The other will unconditionally copy
it.