Change framebuffers such that they capture *cell objects
We're going to start using pointers to more ojects instead of copying
everything all the time so we are more efficient when we do copy (just
the pointer) and use less memory overall. This should further reduce
the burdern on garbage collection.

The change here will mean that we need to be more careful about never
mutating an object after it is created though. Assinging a cell to a
framebuffer should always be done via a new cell object, not by
mutating the existing one. That should match what we're doing now
already, but it will become a hard requirement.