Change the UDP setup for clients
To ensure we can roam effectively when the source network changes, we
need to be bound to a non-specific IP on the client side of the
connection so that an appropriate route to the remote is always
dynamic. Previously, we were Dial()-ing the remote to establish the
client side which always bound the specific IP that was used for the
route to the remote at setup time. Now, we ListenUDP, just like the
server, so we can bind to [::]:0 (later we'll allow forcing ipv4 vs
ipv6 for both ends of the encrypted tunnel). This means that we no
longer need a distinction between client and server in network.Write()
as we always use WriteToUDP and pass in the remote UDP address.