Better protect use of ssh agent sockets
Don't panic when no agent forwarding was requested by validating that
the sock objects aren't nil before using them. Also, bail with an
error if the client gets an agent request but doesn't have a working
local socket. The client shouldn't allow that to happen, but add the
defense anyway.